Move CSharpAsyncVoidMethod test
